+++ lib/lint_analyzer.dart
@@ -1,9 +1,13 @@
 export 'package:dart_code_metrics/src/analyzers/lint_analyzer/lint_analyzer.dart';
 export 'package:dart_code_metrics/src/analyzers/lint_analyzer/lint_config.dart';
+export 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics/models/metric_documentation.dart';
+export 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics/models/metric_value.dart';
 export 'package:dart_code_metrics/src/analyzers/lint_analyzer/metrics/models/metric_value_level.dart';
+export 'package:dart_code_metrics/src/analyzers/lint_analyzer/models/entity_type.dart';
 export 'package:dart_code_metrics/src/analyzers/lint_analyzer/models/issue.dart';
 export 'package:dart_code_metrics/src/analyzers/lint_analyzer/models/lint_file_report.dart';
 export 'package:dart_code_metrics/src/analyzers/lint_analyzer/models/replacement.dart';
+export 'package:dart_code_metrics/src/analyzers/lint_analyzer/models/report.dart';
 export 'package:dart_code_metrics/src/analyzers/lint_analyzer/models/severity.dart';
 export 'package:dart_code_metrics/src/analyzers/lint_analyzer/models/summary_lint_report_record.dart';
 export 'package:dart_code_metrics/src/analyzers/lint_analyzer/models/summary_lint_report_record_status.dart';